+++ test/integration/cluster/cluster_commands_test.rb
@@ -32,6 +32,7 @@ class TestClusterCommands < Minitest::Test
   # ValkeyTests modules without conflicting setup methods
   include ValkeyTests::Bitpos
   include ValkeyTests::Call
+  include ValkeyTests::ClientInfoTag
   include ValkeyTests::GenericCommands
   include ValkeyTests::Scanning
   include ValkeyTests::ScriptingCommands